% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_avg.R
\name{add}
\alias{add}
\alias{times}
\title{Basic arithmetic}
\usage{
add(x, y)

times(x, y)
}
\arguments{
\item{x, y}{numeric vectors. Eqn 
\eqn{a + b} Deqn 
\deqn{a + b} End}
}
\description{
Basic arithmetic
}
