% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_avg.R
\name{moving_avg}
\alias{moving_avg}
\title{A Function to calculate the moving average}
\usage{
moving_avg(v, n = 7)
}
\arguments{
\item{v:}{the vector over which to calculate the moving average.}

\item{n:}{the interval to consider.}
}
\description{
Calculate the moving average of a vector
}
\examples{
moving_avg(1:10, n=7) 
[1] NA NA NA NA NA NA  4  5  6  7
}
\keyword{average}
\keyword{moving}
